Do not depend on the scanner when booting up
This is the first step towards a lexer written in Rebol using PARSE. The
first stage is to have a usable PARSE, which requires the interpreter to
boot up. Instead of creating a boot_array by scanning the rebol code,
(generated) C code is used to create the boot_array directly.
make-rebol-code.r is such a tool that generates C code that creats a
REBOL block.

The second step would be creating a PARSE rule to replace C code in
Scan_UTF8_Manged, and that rule will also be created by C code.