Finer-grained errors from scanner, detect mismatched delimiters
The scanner had a uniform means of delivery, so that particular kinds
of scan errors were not getting distinct identities or arguments.
This is a first cut at changing the approach so that any particular
error could be given a unique identity if needed.

The main motivation was to get better discernment of cases like:

    load "[ abc"
    load "[ abc )"
    load "abc ]"

Only cases like the first can be resolved by more input, so the
usermode REPL will keep allowing more typing when errors of that form
occur, but not the others.  This resolves bugs that were related to
not being able to distinguish cases, in particular of open strings.

This is only a first step, with the intention to allow the scan state
to be processed in a stack to provide a complete list of the tokens
which are outstanding.