Change `scan_state` to just `ss`
Half the lines of code in the scanner referenced `scan_state->xxx`,
which is distracting.  As Do_Core() does by just calling the frame
pointer "f", this helps minimize the noise, making it easier to
read the relevant code if you take for granted that a SCAN_STATE
is there.

Used `ss` instead of just `s` due to common abbreviation of `s` for
REBSER*.