Boot error tolerance patches for REPL skin and Encap
The intended workaround from a prior commit tried to "return blank"
out of a TRAP/WITH handler, which did not import the return via
`<with>` from the enclosing function.  Hence it didn't actually
suppress the error case it was intending, since it returned from the
error handler itself.

Also includes tolerance for if the system/user/rebol path did not
get set, and was BLANK!.  Includes 80-column and other formatting
tweaks for the system/repl initialization.