Improve performance with DEBUG=asserts, including -O2
This does some performance improvement of the debug build in general,
along with making it so that the debug build with asserts w/o symbols
uses the optimizer.  (There's not much reason not to use an optimized
build if it's un-debuggable due to lack of symbols...)

Performance improvements were based on noticing spikes in callgrind,
largely related to function facade macros being called often, and
the bit testing in GET_VAL_FLAG being excessive.  This changes to
just check that GET_VAL_FLAG is on a value whose raw type is > REB_0
and <= REB_MAX...which should catch cell-based ends, trash, and most
non-cell-based ends.