Make GET-ENV return blank instead of void when variable not present
Thinking has evolved on "void", which is not technically an ANY-VALUE!
and can thus be an out-of-band signal for when a value is not present.
Many traditional operations which would use NONE! (now BLANK) for
convenience had been tried out using voids instead to avoid conflating
the blank value with the absence of a value.  Today's compromise leans
toward using the voids in internal routines, and then making wrappers
which "blankify" them for convenience.

Since that is becoming a large assumption, even that GET itself will
return blanks by default when a variable is not set, it makes sense
for GET-ENV to do this as well.