Plain GET returns BLANK! for not set variables, GET-VALUE will error
This normalizes the GET and SET natives, making the code run for a
GET/SET of a single value go through the same basic code path as
that for a block.  Hence you can do `set [x x/y x/y/z] [1 2 3]` for
example.  It also adds checks to avoid GROUP!s in paths, which could
have side effects other than GET-ting or SET-ting, e.g.
`get 'x/(print "hi" 'y)`...this could lead to confusing behavior and
is better done other ways such as with REDUCE.

The SET and GET behavior for objects from blocks, which has always been
sketchy, is moved to usermode and legacy.  VARS-OF would be a better
name for getting the variables of an object, and it could be that SET
is an action which would be overridden for user defined types anyway.

A big change to try and align with the convenience differences e.g.
between SELECT and SELECT* is the "blankification" of GET's result by
default.  This helps reduce the disruption of `block/VALUE` no longer
giving back blank, because `get 'block/VALUE` looks about as nice and
literate and has less at stake--due to not being able to dispatch
functions etc.