Fix stack labeling of functions dispatched from mid-path
Functions which were dispatched from the middle of a path (e.g. a
member of an object) were being labeled in the stack frame as
anonymous, despite a name for them being available, if they did not
have any refinements.  This was because the label check was being
done before the path dispatch step, instead of after it.