Stack-based REBFRM* can act as specifiers, kill VALUE_FLAG_RELATIVE
This is a significant commit which introduces the new idea of being
able to put pointers to stack quantities inside REBVALs.

That can't be universally safe, since some REBVALs live in arrays or
objects.  The stack level where a stack-based pointer originates from
might get popped while the pointer is still referenced.  Hence it must
be possible to tell by looking at a cell what its duration is...and if
a raw pointer held in a source cell can't be put into a destination
cell safely, it must be reified into a heap pointer and managed by GC.

Included in the design is the concept of deciding the nature of a
binding (relative vs. specific) entirely by examining the flags of the
REBSER node to which a value is bound.  The binding is also never set
to NULL...the global empty array series is used, which allows to test
series bits without a NULL check.

Although this means a REBFRM* does not need to be "reified" immediately
on calling a user function (as it did before), the conservative logic
for reification during `Move_Value()` is only a prototype.  Any leaked
words to *any* cell will cause a reification at the moment--just to be
on the safe side.  But even this should be an improvement, as native
functions still will never reify (since they don't leak bound words
referring to their arg locals) and not all user functions involve
moving words bound to their arguments or locals into new REBVAL slots
for those words.

This pioneers generic approaches to using stack values to augment the
interpretation of REBVALs without the overhead of dynamic allocation
or creating entities that will tax the GC...so-called "virtual binding"