Patch for Windows Unicode support in console
When Windows is redirecting its input to a file or socket, then it
should use UTF-8 bytes.  But when communicating specifically with a
Windows console, UTF-8 cannot be used...the API for the Windows
console predates UTF-8 (and is based on some kind of UCS-2 or UTF-16
or whatever the distinction is there).

There was code to detect this condition but it was written incorrectly,
assuming that the windows console was FILE_TYPE_UNKNOWN when it's
actually FILE_TYPE_CHAR.  It may be more subtle than that, having to
do with things like "printer redirection", but this seems to address
the problem.