Make WHAT-DIR return copy of system variable
The OS has a notion of "current directory", but Ren-C extends this
concept to be potentially a URL, so that WHAT-DIR requests can come
back reflecting the place where a network script is running.  This
means that the OS's internal directory notion isn't the only place
the current directory can be coming from.

Because of this, a system options variable for the current directory
is kept.  However, on WHAT-DIR calls if it is out of sync, it is
refreshed from the host's notion of what the current directory is.

The internal variable was being returned without being copied, when
routines like CLEAN-PATH expected it to be mutable.  This copies the
sequence before returning it.

Includes light improvement of CLEAN-PATH, line ending fixup in the
detab-file, and a fix for UNUSED() variables from the last commit.