Fix DNS reverse lookup when given a string
R3-Alpha's DNS actor was incorrectly using Scan_Tuple on the data of
a tuple, instead of on a string to convert it to a tuple.

https://github.com/rebol/rebol/blob/25033f897b2bd466068d7663563cd3ff64740b94/src/core/p-dns.c#L69

The interface for Scan_Tuple is supposed to be for *making* a tuple out
of a string, not making a tuple out of a tuple's data.

This is a believed fix for reverse DNS lookup.