Fix import 'test by SET/PAD instead of plain SET in LOAD-MODULE
To avoid a potentially bad variance in behavior from `set [a b] c`,
where c may be either a block or non-block value, the behavior of
allowing a block on the set side to be paired with a non-block was
removed.  The chaining behavior of tolerating a blank in and blank
out was kept so long as you used /PAD, so that it becomes more clear
at the callsite when a set is doing this.

The main user of blank chaining in SET of BLOCK! was BrianH in the
module code.  Some instances were not being exercised by the standard
code paths, and IMPORT 'TEST would get an error.