Fix Ctrl-C handling in Windows console IO
When ReadConsole is called, a Ctrl-C during the read will end up
returning a zero length wide char string.  It turns out that
WideCharToMultibyte fails if cchWideChar is 0, so this was causing
a Ctrl-C to raise an I/O error...instead of just returning an empty
string and then having the Ctrl-C handler routine run.

As a secondary problem making the first harder to find, the latent
nature of Ctrl-C processing was causing a panic to be interrupted while
the crash was trying to report itself.  This gets around that by
unregistering the Ctrl-C handler before calling panic outside of any
PUSH_UNHALTABLE_TRAP.