Error message when voids are hit in path dispatch
The unification of PICK* and pathing meant that voids were arising
more often in path dispatch, yet there is no dispatcher for void.  To
make one would require making the Path_Dispatch table one longer to
include REB_MAX (the type used for voids).  This just puts in an IF
statement for now.