Add /ONLY refinement to SET, permit single value to block sets
Allowing `set [a b] 10` was removed, because it seemed to have a "bad"
polymorphism that if `set [a b] value` was used, the value could be
either a block or a non-block...giving SET different semantics.

It turned out the feature was used more widely than expected, because
of code that expected to be able to propagate blanks e.g. a failed
result from FIND:

    all [
        set [a b] find data 'thing
        ...
    ]

So a more limited version was added in, that allowed BLANK! when
SET/PAD was used.

Later it was noticed that Red had made an /ONLY refinement and was
treating SET as if it were in the same family as APPEND and other
operations that treat blocks differently from non-blocks.  Being
prescriptive and thinking that is unsafe could be taken care of with
an adaptation that checked the types, so this embraces the /ONLY
refinement and allows the tricky behavior in the primitive.