Add field selectors by pointer to spelling of word for contexts
While most hardcoded selections in contexts in the core are done by
integer (due to precalculation of positions in system objects),
extensions and other applications might want a more convenient way of
selecting by word spelling.  This adds a couple of convenience
functions to do that.

Also a couple of CMake comment tweaks and deletion of the evil old
"last N series that the GC won't consider" thing that was in R3-Alpha.