Change lingering SET_XXX functions to Init_XXX functions
For discussion on this, see:

https://github.com/metaeducation/ren-c/pull/457

R3-Alpha had functions like VAL_SET_BLOCK() or VAL_SET_WORD(), which
were confusing to those trying to read the code given the SET operation
and the multiple meanings this might have.  Over time, Ren-C shifted
this to VAL_INIT_XXX(), eventually dropping the VAL and just going to
plain Init_XXX().

Though this was done for the complex types, the simpler types had
stayed with names like SET_INTEGER().  This brings them in sync with
Init_Integer() naming, etc.