Make code, type, id blank for "user" errors - release build file/line
The RE_USER error code was changed to be a very large number (MAXINT),
which was a visible disruption if a user were to trap the error.  It
was also not a number anyone would type in.

What makes more sense in a system which has the ability to express out
of band values of "no code" is to use the standard way of saying "not
there", e.g. a BLANK!.  This is easier to test for, but required some
changes to the molding and other mechanisms that expected the fields to
be specific types.

RE_USER is kept and properly accomodated to build errors using the
Error() function, which means the string argument is actually the
message... not an argument.  That fixed a bug which prevented the
`fail ("of a string");` from working correctly.

The error catalog and behavior was unusually complex and difficult to
understand when first looking at R3-Alpha, yet a disproportionate
number of tests were written for it.  In order to document it and keep
it working, while pursuing the idea that errors would not have 3 args
named "arg1", "arg2", and "arg3", it was beefed up with asserts in the
early days of Ren-C.  While this informed many findings about how
series and contexts worked to improve other aspects of the code, the
compatibility with arg1/arg2/arg3 is something that could now be
hooked in a legacy compatibility in usermode in MAKE, so the relatively
complex C code can be dropped.

This also goes ahead and takes the `__FILE__` and `__LINE__` debug
feature for fail()s that originate in C code and legitimizes a file
and line field in the error in the release build.  That's due to a
desire to continue developing the feature connecting usermode series
back to the source using the unused ->misc and ->link fields in
those series.  (That feature is not completely ready yet, but is
under development).