Correct bug in prior fix for frame labeling, add file and line debug
A previous fix to collecting the frame labels was getting a label but
it was incorrect...winding up being the name of the last refinement.
This fixes that.

It also extends the debugging information built into frames to account
for any file and line information which is tagged onto a source array
at a given stack level.  As that feature improves, this should help
locate approximately where in a Rebol source file the given code is
running from.

The names for the fields that are debug only are also changed to be
suffixed with _debug, to make it clearer in code that uses them that
these values are not part of the release build.  This reduces the
need to emphasize that with an additional #if !defined(NDEBUG) that
surrounds an assert.