Re-adopt convention of array heads passed as [] not * as args
In C, a function arg of `FOO foo[]` and `FOO *foo` are mechanically
equivalent.  However, the former helps cue readers to know that the
pointer will be incremented and indexed, as opposed to being merely
a single location to be written.

This cueing was added at one point in the source, but removed when the
idea of using a class for a REBVAL pointer (e.g. `REBVAL(*)`) was
being experimented with.  That experiment was ultimately abandoned,
so there's not really any good reason not to re-instate the convention.

Includes fix for whitespace in %host-start.r