Fix aggressive warning in gcc 4.9.2 in %f-deci.c
Some old math code took a const parameter which it knew to not be 0
and used it to index an array as `d[m - 1]`.  Since the value was
listed as const, the compiler in -O2 was able to do analysis that
decided a zero value would be a problem.  It's somewhat presumptuous
to assume the callee and caller do not have a rule that there is no
zero, but still better to cover the case...so this tweaks the code
to make sure a branch handles zero, and also includes an assert that
m is not 0 on that branch.