Move init code of crypt extension to its own file
This introduces another way to embed the init script in the extension
and makes it compressed. The init script will be processed by
make-ext-init.r, and then be written to a C file.

load-extension tells if the data is compressed or not by its type:
compressed is in the binary! form and uncompressed is in the string!
form.