Standardize OS_Browse API to boolean result
The return result from the host kit service API for opening a web
browser was sometimes an integer expected to be 0, and sometimes a
boolean.  If boolean, then TRUE would be interpreted as a 1 and hence
an error to report, but the browser would launch anyway.

Ultimately, the goal is that services in host kit will be extensions
and able to raise their own specific fail() calls, rather than piping
information through some integer-based interface.  So the shortest path
to that future is to pick the narrowest common denominator result code
of TRUE for browser worked and FALSE for not.