* Make SET have /PAD semantics by default, add SET/SOME refinement
R3-Alpha had a refinement /PAD which suggested that if SET was being
used with a block of values, if the block was too short for the number
of words being set, it would blank them.  This commit makes that the
default behavior, and adopts /SOME (a refinement introduced by Red) to
indicate that blanks in the supplied block or slots past the end do
not affect the values of the indicated variables.