Add a function to retrieve executable path
arg[0] is not reliable to retrieve the executable path, because when the
executable in $PATH, arg[0] could be just the file name, instead of a
full/relative path.

This adds a platform specific function to do this. See:
https://stackoverflow.com/questions/1023306/finding-current-executables-path-without-proc-self-exe

For platforms this function is not implemented, it will fall back to
argv[0].