Make it possible to explicitly set optimization level
if OPTIMIZE= is passed to makefile.boot, it will overwrite the
optimization level set by DEBUG=, unless it's "auto", so user can do:

$make -f makefile.boot DEBUG=sanitize OPTIMIZE=3

to have optmized build with address sanitizer.

Setting it to "auto" has the same effect as it being omitted.

This also changes the optimization level of bootstrapping r3-make to 2,
because it's so slow that travis times out sometimes. Because we don't
debug these binaries (they are not saved), "-O0" doesn't mean much, and
the optimized version is really the one we should test with.