Move CALL and BROWSE into an extension
The division of labor in R3-Alpha for code that required a different
implementation on various OSes was to call OS_XXX functions, that were
linked separately as part of the host's implementation.  These
functions were intended to be C code that did not rely on Rebol types
(e.g. any string parameters to these services would have to be
transformed into the OS-native representation) and memory had to
be explicitly managed by contract, vs leveraging Rebol series.

The consequence of this division was that there would be a native
for a function like CALL, which would need to translate from Rebol
types into non-Rebol structures and flags, which generated overhead...
and then the routine receiving the unpacked data would have to give
back its results to be turned back into Rebol again.

This begins a transition to where CALL itself lives in an extension
which is itself able to use the internal Rebol API.  Not only does this
permit builds to omit CALL (e.g. emscripten would not have had any
underlying OS_XXX function for process spawning), but it means the
code can be much more straightforward.

This only does the conversion to an extension--it does not yet make any
significant changes to the code.  Being done as a separate commit to
see if Travis can build it.