* Move BROWSE into userspace, built upon CALL
Previously, the CALL mechanics were reused by way of a shared C routine
in the BROWSE native.  This commit alters the form of reuse to actually
use the CALL native directly, putting BROWSE into userspace.

Because some C code may be necessary to detect what browser(s) are on
the system (e.g. by way of registry key access), this moves just that
part into a native called GET-OS-BROWSERS.

Includes a fix for a bug introduced in Linux builds that was
introduced while extracting to a module.