Begin removing redundancy and using internal APIs in CALL
This starts taking steps to remove the redundancy between the code
implementing CALL on Posix and on Windows.  The various flags or
parameters for translating input types are turned into just basic
checks of VAL_TYPE(ARG(...)) or of the refinement flags.

As an interim trick, the native's frame_ pointer is passed in to give
the formerly independent OS_Create_Process() functions access to the
parameters, instead of having to find ways to coerce them into non
Rebol types.

Once the OS_Create_Process() functions are able to operation on actual
Rebol binaries and strings directly, they'll be able to progressively
append content to the output--rather than needing to do a large
non-Rebol allocation that is returned all at once.