Assgin canonical form to extension words
The intension of the existance of the array canons was to save the their
canonical form and make the comparison easier. However,
Intern_UTF8_Managed might not return a canonical form of a string (when
it's canonical form already exists, and different from the word in
question), which makes direct comparison fail.