Add "falls through" comments
GCC 7 will have -Wimplicit-fallthrough turned on by -Wextra. There are a
number of ways to inform the compiler that falling through is intended.
This commit just uses the comment to do that, which is the most
conservative (least destructive) way: if the compiler doesn't recognize
it, nothing's gonna be broken.