Move FFI to extension, including STRUCT! callbacks
UNFINISHED: Does not support molding or comparison of STRUCT!

When Ren-C was being extracted from Atronix's repository, it was known
that it would not be desired that all builds include FFI.  It was a lot
of code, and not code that makes sense on all platforms (e.g.
emscripten).  Yet it had fairly deep hooks into the system, including
the memory pools and core function handling.

Instead of #ifdef'ing out the FFI code and letting it atrophy, it was
kept building in Ren-C over time--without requiring being linked to
the actual FFI library.  This was done by way of a set of "FFI stubs",
that would give errors if the code was run without the FFI, yet would
still compile.  This kept FFI functionality on the radar--but also
meant all the code for the FFI interface was in all builds (minus the
size of the libffi library at link time)

Over time, many of the deep hooks were factored out--especially those
affecting the garbage collector and memory allocation code.  This
commit capitalizes on that transition to move the Rebol-based FFI
processing code out of the core, making it fully optional as an
extension.

However, some ideas for extensions--such as user defined types--aren't
actually ready yet.  Even though STRUCT! was rethought as something
that would be easier to do with user defined types than previously,
things like path dispatch or MAKE haven't been designed for user
defined types.

This creates a minimal set of hacky hooks for allowing the Reb_Struct
to be defined inside of the core, but the behaviors like path dispatch
registered from within an extension.  While not perfect, it does mean
that it's possible to lighten up a build size by omitting the
extension, or by having it dynamically loaded as a DLL.