Add a native to set C break point
The debug version of the interpreter keeps tracking of a do_count, and
can raise a break signal at a given do_count point before evaluating the
expression. This is very useful for debugging the evaluator, however,
it can only break at a hardcoded point, which means it needs to be
recompiled after changing; moreover, it's very hard to find out the
do_count to break.

This commit adds a rebol function to set such a break point, no
recompilation is needed, and multiple breaking points can be set. When
the function is called, it will break at next point, in next x points
by using /skip, or at an certain point by using /at

e.g.
    c-break-debug
    a: 1

will make it break in Do_Core right before evaluating "a: 1"