Rework on make-make.r
Now make-make.r is based on rebmake.r, which can generate Makefiles for
GNU make and Microsoft NMake, and supports different compilers
(currently, gcc and CL are supported). It can also run commands without
generating a makefile at all.

The goal of rebmake is to be a generic makefile/project generator.
make-make.r is the actual driver for bulding the interpreter and its
extensions.

The motivation to rework on the building system is to have a single
system that:
1) user can choose which extension/modules to build, and developers can
easily define module/extension/library dependencies, and how the
extension should be built
2) can generate IDE project file, being able to debug the interpreter in
Visual Studio is invaluable when developing on Windows.

Currently there are two parallel building systems: make-make.r and
CMake. Either of them is good enough for our purpose: make-make.r is
simple, but very limited; cmake is good at generating IDE project files,
but not well integrated with the system (not eays to define dependency,
or module/extension building procedure)