Support platform dependant module natives
This commit adds a new block to the native spec: "platforms: []", the
block can include os-base (in which case all platforms based on it will
be supported) or os-name, e.g [android linux-x64] to support only
android and linux-x64

When the "platforms" block is missing, it's assumed to support all
platforms.

Adding "platforms" block only makes it visuable in the module for these
platforms, to make it really work, the C source has to be proteced by
"#if" and "#endif". So there's repeatation in "platforms:" and the #if
directive. Ideally, this should be elimininated, but as they are
processed by different tools (make-ext-natives.r and compiler), more
work is needed.