Un-methodize devreq size so it's just req_size in the Device
A port has a certain fixed-size amount of state, which could contain
arbitrary C pointers or data structures.  The ability to change the
size of this state at runtime was removed, as a step toward making the
devices store their state (strings, etc.) in Rebol-readable values.
To make things more specific, each device was allowed to have its own
specialized structure; but the size was acquired via a method call.

This simplifies that by just passing the sizeof() at the time of the
registering of the device.