Rename SET_UNREADABLE_BLANK to Init_Unreadable_Blank
The fact that there's only an unreadableness property in the debug
build raised a question of if calling it "Init_Unreadable_Blank" in
the release build made sense.  That suggests some kind of property that
is not checked at runtime without a qualifier.

Although Init_Blank_Unreadable_If_Debug() is verbose, it *might* help
convey a better understanding of what's actually going on.  But it
can also be confusing, because it's not clear if there is no operation
in the release build at all.  So is it (Init_Blank_Unreadable) If_Debug
or (Init_Blank) Unreadable_If_Debug.  Maybe just making matters worse.

Given that you're never supposed to write to these in the release
build either, the simpler and easier to read Init_Unreadable_Blank()
is probably the better option, so this goes ahead and does that.