Set_Stack_Limit during PUSH_UNHALTABLE_TRAP
This is a trick which can be used to allow several different threads to
work with the same evaluator instance.  The rule is that each thread
must PUSH_UNHALTABLE_TRAP, and when it does, that will reset the
stack limit used by the test in C_STACK_OVERFLOWING.

This is an attempt to be minimally invasive, given that switching
threads while keeping the same frame stack of Reb_States should not
longjmp in case of a fail() to another stack; hence pushing a trap
is *always* necessary.  But it's kludgey compared to a real solution,
which would have to be a future endeavor.

Also corrects a >= 0 comparison on an unsigned value from the previous
commit.