Fix stack overflow when molding cyclical structures
There was an array of values used to notice when a mold was recursing,
that was used to stop the molding and output a "..." instead.  At some
point when the logic behind having managed vs. unmanaged series was
being developed, the value being tested for in this array was not the
array itself, but the first value in the array.  Hence it would only
stop recursions if the array recursed in its first element.

This commit simplifies the code, and dodges an issue about molding an
unmanaged series (e.g. for debugging) by using a series of raw pointers
instead of values for the recursion.

Also, this commit adds a recursive case to the test suite.