Use empty blocks instead of BLANK! for %systems.r
In R3-Alpha, an underscore is just a WORD!.  A compatibility hack makes
it evaluate to a #[none], but that hack only works in evaluative
contexts.

There's no particularly compelling reason why an empty block couldn't
be used in the %systems.r table, so this switches it...rather than try
and come up with a workaround.