Update FUNC, FUNCTION, UNSPACED, etc. in R3-Alpha compatibility
The use of tags like `<local>` or `<with>` is now rather pervasive,
and so it's worth it to go ahead and emulate these when running under
R3-Alpha instead of trying to force people to use the old conventions.

Given that a lot of code would have to be changed at this point if the
convention changed, this also goes ahead and changes things to use
the local tag outside of the code that was using it for compatibility.

Also implements the UNSPACED and SPACED logic of being able to omit
voids and blanks, which makes it more pleasant to work with IF or
other conditionals inside of them.