Add -rdynamic link option to r3
When extensions are built as dynamic loadable libraries, they need to
import symbols from the main executable, and these symbols have to be
exported by the executable. With ELF and GCC, this is done by
-rdynamic:

"Pass the flag -export-dynamic to the ELF linker, on targets that
support it. This instructs the linker to add all symbols, not only used
ones, to the dynamic symbol table. This option is needed for some uses
of dlopen or to allow obtaining backtraces from within a program."