Move ext-dynamic-objs after dynamic-libs
So that the object files will be build as dependencies of a dynamic
library, and proper parameters (-fPIC) will be passed in.

Because every target will only generate one rule, if the object file
was specified before a dynamic library, it has no idea of the dependency
during the command deduction process, and thus no "-fPIC" parameter is
added. and when it gets to the point to generate the dynamic library,
(rules of) object files have already been generated, and will be
skipped. Thus, "-fPIC" will be missing from the command.