Disable RETURN in FOR-EACH-RECORD vs. giving comment warnings
Rather than warning that RETURN doesn't work the same way in R3-Alpha
that it does in Ren-C on the callsites to FOR-EACH-RECORD, just go
ahead and make it bound to a function that will raise an error.

At the same time, use the creation of this new context and binding to
make the bound variable not overwrite any previous definition, as
FOR-EACH and other constructs do in R3-Alpha (though Red has given up
on that).