Make SET not tolerate BLANK!, use BAR! for dummy targets
The feature of tolerating blanks was designed to permit things like:

    set [a _ b] [1 2 3]

Such that a would be set to 1, b to 3, and not have to create a dummy
variable to assign 2 to.  But it's too easy to wind up with something
like `set in obj 'word value`, where the IN returns a blank and the
SET has no error.

This takes the somewhat obscure feature and does it with BAR! instead

    set [a | b] [1 2 3]

...where previously, BAR! would be a noop, but only if there was a
BAR! on the right side.  So it's still possible to leave out a target,
just harder to do accidentally.