Add '--as-needed' link flag to dynamic extensions
All of the libraries that the main app depends on will be passed to the
extension when they are built as dynamic libraries, so that it just
works the same way as it's built into the main executable. But all of
these libraries will be linked into the extension (as shown by 'ldd
-u'), even though they don't directly depend on, '--as-needed' is a
linker flag to prevent such unneeded linking.