Rework on the command assembly logic
Most of the time, the command (format) is determined solely by the
target platform, but not in all cases: mingw-make if launched from cmd,
would expect commands in windows format, and while launched from bash,
it would them in POSIX format.

This change separates the command data and its assembly, which makes it
more flexible: A platform can have a default assembler and the target
generator (e.g. makefile) can overwrite that, which enables the
execution target to cross-compile.