Allow DO of a variadic parameter
This allows a hard-quoted variadic parameter to be used with DO.

    foo: func [:x [any-value! <...>] <local> pos] [
        do/next x 'pos
    ]

    >> foo x: 1 + 2
    == 3

This is then applied to the DUMP routine, so that if a SET-WORD! comes
immediately after dump (e.g. `dump x: 1 + 2`) it senses that, and
assumes you would like to evaluate a single DO/NEXT and then label
the result with the set-word it sensed.

    >> x: 10

    >> dump x 3 + 4 ;-- ordinary case, just looks at `x`
    x: => 10
    == 7

    >> dump x: 1 + 2 ;-- newfangled case, leverages DO
    x: => 3

DUMP of a BAR! now disables the dump.  Hence instead of:

    longvariablename: 1 + 2
    dump longvariablename

You can save on typing with:

    dump longvariablename: 1 + 2

And instead of disabling it with

    longvariablename: 1 + 2
    ;dump longvariablename

You can simply say:

    dump | longvariablename: 1 + 2

(Neither is a very good practice to commit when you are done debugging;
leaving in the dump has a small runtime cost, and committing commented
out code is a generally poor practice.)

The precise rules of how typing interacts with this function are not
quite yet established.  And there's some other code in a pending commit
that mucks around with how the parameter tying works, so it's left TBD.

TO BE SAFE: assume in the end that it will only be allowed on hard
quoted ANY-VALUE! variadic parameters.  This is because that's
theoretically enough permission that the routine could have simulated
the evaluator by going element by element and re-implementing it.

(As another idea, it might be legal if you say `foo 1 + 2` to have foo
work if its argument is a hard quoted `:arg [word! integer!]`.  That
might be technically hard to enforce, and it also might be dumb...so
the idea might be that if you have typing and you DO the parameter,
it's implied you meant to check the result of the DO.)