Convert file paths to local format
Most of the windows tools can work with '/' as the path delimiter, but
some are not. To be consistent, just do conversion in all cases.

In some cases, they should actually be in the target OS format, because
that might be different from local system, but currently, Rebol doesn't
have functions to do that, and probably not many people are doing
cross-compiling anyway, so ...