Simplify FAIL SPACED and ALL [LOGIC? ...] cases
It is now pretty much firm that PRINT and FAIL are to be based on the
idea that they use some form of SPACED logic intrinsically.  Not only
is this consistent with history, extended study couldn't come up with
a cleaner or better alternative, than asking those who did not need
this default to use PRINT UNSPACED or FAIL UNSPACED.

Also, LOGIC!-testing is simpler when compared against the direct
values vs things like `all [logic? debug debug]`.  That case is more
simply handled as `debug = true`.